app/requirements.txt
asgiref==3.8.1
dj-database-url==2.3.0
Django==5.1.5
numpy==2.2.2
pandas==2.2.3
python-dateutil==2.9.0.post0
python-slugify==8.0.4
pytz==2025.1
six==1.17.0
sqlparse==0.5.3
text-unidecode==1.3
typing_extensions==4.12.2
tzdata==2025.1
pillow==11.1.0
psycopg2-binary==2.9.10
gunicorn==23.0.0
openpyxl==3.1.5
xlrd==2.0.1
aiohttp==3.11.18
BeautifulSoup4
fuzzywuzzy
requests
python-Levenshtein
